move test to test folder
